Monster stats evolve with level Chest done Plank is a rare loot

+++ ProjetRecode/src/Model/Damageable.java
@@ -1,5 +1,5 @@
 package Model;
 
 public interface Damageable {
-	public void sufferDamage(int damage);
+	public void sufferDamage(float damage);
 }
